+++ apps/decodex/src/recovery/tests.rs
@@ -16,17 +16,10 @@ use crate::{
 	recovery::{
 		GHOST_LANE_CLEANUP_EVENT, GHOST_LANE_TERMINAL_STATUS,
 		MCP_TEST_FIXTURE_GHOST_LANE_CLASSIFICATION, RecoveryRuntimeMutationPolicy,
-		active_recovery_tracker_backoff_message, append_review_handoff_adopt_private_event,
-		append_review_handoff_rebind_private_event, apply_ghost_lane_cleanup,
-		apply_ghost_lane_live_status_blockers_with_tracker, current_timestamp,
-		diagnose_all_retained_review_worktrees_with_tracker, diagnose_ghost_lanes,
-		diagnose_ghost_lanes_read_only, diagnose_issue_with_tracker, diagnostic_binding,
-		ensure_ghost_lane_live_status_allows_cleanup_with_tracker, manual_adopt_run_id,
+		active_recovery_tracker_backoff_message, current_timestamp,
+		diagnose_all_retained_review_worktrees_with_tracker, diagnose_issue_with_tracker,
 		remember_recovery_tracker_backoff_message, timestamp_after_seconds,
-		validate_adopt_existing_worktree_mapping, validate_adopt_issue_state_for_policy,
-		validate_adopt_landing_state, validate_existing_handoff_refresh,
-		validate_rebind_issue_state_for_policy, worktree_blocking_status_lines,
-		write_review_lifecycle_markers_with_rollback,
+		worktree_blocking_status_lines,
 	},
 	state::{ProtocolActivitySummary, StateStore, WorktreeMapping},
 	test_support,
