+++ apps/decodex/src/recovery/tests/stale_active/release/preflight.rs
@@ -1,0 +1,279 @@
+use std::fs;
+
+use tempfile::TempDir;
+
+use crate::{
+	recovery::{
+		RecoveryRuntimeMutationPolicy,
+		tests::{self, GhostLaneTestTracker},
+	},
+	state::ReviewPolicyCheckpointInput,
+	tracker,
+};
+
+#[test]
+fn stale_active_release_preflight_rejects_worktree_progress_after_diagnosis() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let context = tests::sample_recovery_context(
+		&temp_dir,
+		RecoveryRuntimeMutationPolicy::AllowRuntimeWrites,
+	);
+	let active_label = tracker::automation_active_label(context.config.service_id());
+	let queue_label = tracker::automation_queue_label(context.config.service_id());
+	let worktree_path = context.config.worktree_root().join("PUB-1626");
+	let mut issue = tests::sample_issue_with_labels("Todo", &[active_label, queue_label]);
+
+	issue.identifier = String::from("PUB-1626");
+
+	tests::init_clean_git_repo_with_remote_default(&worktree_path, "x/pubfi-pub-1626");
+
+	context
+		.state_store
+		.record_run_attempt("run-1626", &issue.id, 1, "running")
+		.expect("run attempt should record");
+	context
+		.state_store
+		.upsert_worktree(
+			context.config.service_id(),
+			&issue.id,
+			"x/pubfi-pub-1626",
+			&worktree_path.display().to_string(),
+		)
+		.expect("worktree mapping should record");
+
+	let tracker = GhostLaneTestTracker::with_issues(vec![issue.clone()]);
+	let mut diagnostics = super::diagnose_stale_active_issues(
+		context.config.service_id(),
+		&context.workflow,
+		context.config.worktree_root(),
+		&context.state_store,
+		&tracker,
+		Some("PUB-1626"),
+		RecoveryRuntimeMutationPolicy::AllowRuntimeWrites,
+	)
+	.expect("stale active diagnosis should run");
+	let diagnostic = diagnostics.pop().expect("diagnostic should exist");
+
+	assert!(diagnostic.recoverable());
+
+	fs::write(worktree_path.join("late_progress.rs"), "fn late_progress() {}\n")
+		.expect("late untracked progress should write");
+
+	let error = super::preflight_stale_active_worktree_cleanup(&context.state_store, &diagnostic)
+		.expect_err("preflight should reject late retained progress");
+
+	assert!(
+		error.to_string().contains("retained worktree changes appeared before cleanup"),
+		"unexpected preflight error: {error:?}"
+	);
+}
+
+#[test]
+fn stale_active_release_revalidates_late_default_worktree_progress_without_mapping() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let context = tests::sample_recovery_context(
+		&temp_dir,
+		RecoveryRuntimeMutationPolicy::AllowRuntimeWrites,
+	);
+	let active_label = tracker::automation_active_label(context.config.service_id());
+	let queue_label = tracker::automation_queue_label(context.config.service_id());
+	let default_worktree_path = context.config.worktree_root().join("PUB-1626");
+	let mut issue = tests::sample_issue_with_labels("Todo", &[active_label.clone(), queue_label]);
+
+	issue.identifier = String::from("PUB-1626");
+
+	context
+		.state_store
+		.record_run_attempt("run-1626", &issue.id, 1, "running")
+		.expect("run attempt should record");
+
+	let tracker = GhostLaneTestTracker::with_issues(vec![issue.clone()]);
+	let mut diagnostics = super::diagnose_stale_active_issues(
+		context.config.service_id(),
+		&context.workflow,
+		context.config.worktree_root(),
+		&context.state_store,
+		&tracker,
+		Some("PUB-1626"),
+		RecoveryRuntimeMutationPolicy::AllowRuntimeWrites,
+	)
+	.expect("stale active diagnosis should run");
+	let diagnostic = diagnostics.pop().expect("diagnostic should exist");
+
+	assert!(diagnostic.recoverable());
+
+	tests::init_git_repo(&default_worktree_path);
+	fs::write(default_worktree_path.join("late_default_progress.rs"), "fn late() {}\n")
+		.expect("late default progress should write");
+
+	let error = super::apply_stale_active_release_with_tracker(
+		&tracker,
+		&context.config,
+		&context.workflow,
+		&context.state_store,
+		&diagnostic,
+	)
+	.expect_err("late default worktree progress should block release");
+	let run = context
+		.state_store
+		.run_attempt("run-1626")
+		.expect("run attempt should read")
+		.expect("run should exist");
+
+	assert!(
+		error.to_string().contains("safety inspection changed before apply"),
+		"unexpected release error: {error:?}"
+	);
+	assert_eq!(run.status(), "running");
+	assert!(tracker.label_removals.borrow().is_empty());
+}
+
+#[test]
+fn stale_active_release_revalidates_late_run_lease_before_mutation() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let context = tests::sample_recovery_context(
+		&temp_dir,
+		RecoveryRuntimeMutationPolicy::AllowRuntimeWrites,
+	);
+	let active_label = tracker::automation_active_label(context.config.service_id());
+	let queue_label = tracker::automation_queue_label(context.config.service_id());
+	let mut issue = tests::sample_issue_with_labels("Todo", &[active_label.clone(), queue_label]);
+
+	issue.identifier = String::from("PUB-1626");
+
+	context
+		.state_store
+		.record_run_attempt("run-1626", &issue.id, 1, "running")
+		.expect("run attempt should record");
+	context
+		.state_store
+		.upsert_worktree(
+			context.config.service_id(),
+			&issue.id,
+			"x/pubfi-pub-1626",
+			&context.config.worktree_root().join("PUB-1626").display().to_string(),
+		)
+		.expect("worktree mapping should record");
+
+	let tracker = GhostLaneTestTracker::with_issues(vec![issue.clone()]);
+	let mut diagnostics = super::diagnose_stale_active_issues(
+		context.config.service_id(),
+		&context.workflow,
+		context.config.worktree_root(),
+		&context.state_store,
+		&tracker,
+		Some("PUB-1626"),
+		RecoveryRuntimeMutationPolicy::ReadOnly,
+	)
+	.expect("stale active diagnosis should run");
+	let diagnostic = diagnostics.pop().expect("diagnostic should exist");
+
+	assert!(diagnostic.recoverable());
+
+	context
+		.state_store
+		.upsert_lease(context.config.service_id(), &issue.id, "run-1626", "In Progress")
+		.expect("late lease should record");
+
+	let error = super::apply_stale_active_release_with_tracker(
+		&tracker,
+		&context.config,
+		&context.workflow,
+		&context.state_store,
+		&diagnostic,
+	)
+	.expect_err("late run lease should block release");
+	let run = context
+		.state_store
+		.run_attempt("run-1626")
+		.expect("run attempt should read")
+		.expect("run should exist");
+
+	assert!(
+		error.to_string().contains("safety inspection changed before apply"),
+		"unexpected release error: {error:?}"
+	);
+	assert_eq!(run.status(), "running");
+	assert!(tracker.label_removals.borrow().is_empty());
+}
+
+#[test]
+fn stale_active_release_revalidates_late_review_policy_before_mutation() {
+	let temp_dir = TempDir::new().expect("tempdir should create");
+	let context = tests::sample_recovery_context(
+		&temp_dir,
+		RecoveryRuntimeMutationPolicy::AllowRuntimeWrites,
+	);
+	let active_label = tracker::automation_active_label(context.config.service_id());
+	let queue_label = tracker::automation_queue_label(context.config.service_id());
+	let mut issue = tests::sample_issue_with_labels("Todo", &[active_label, queue_label]);
+
+	issue.identifier = String::from("PUB-1626");
+
+	context
+		.state_store
+		.record_run_attempt("run-1626", &issue.id, 1, "running")
+		.expect("run attempt should record");
+	context
+		.state_store
+		.upsert_worktree(
+			context.config.service_id(),
+			&issue.id,
+			"x/pubfi-pub-1626",
+			&context.config.worktree_root().join("PUB-1626").display().to_string(),
+		)
+		.expect("worktree mapping should record");
+
+	let tracker = GhostLaneTestTracker::with_issues(vec![issue.clone()]);
+	let mut diagnostics = super::diagnose_stale_active_issues(
+		context.config.service_id(),
+		&context.workflow,
+		context.config.worktree_root(),
+		&context.state_store,
+		&tracker,
+		Some("PUB-1626"),
+		RecoveryRuntimeMutationPolicy::ReadOnly,
+	)
+	.expect("stale active diagnosis should run");
+	let diagnostic = diagnostics.pop().expect("diagnostic should exist");
+
+	assert!(diagnostic.recoverable());
+
+	context
+		.state_store
+		.upsert_review_policy_checkpoint(ReviewPolicyCheckpointInput {
+			project_id: context.config.service_id(),
+			issue_id: &issue.id,
+			run_id: "run-1626",
+			attempt_number: 1,
+			phase: "handoff",
+			review_level: "normal",
+			status: "clean",
+			head_sha: "2222222222222222222222222222222222222222",
+			nonclean_rounds: 0,
+			details_json: "{}",
+		})
+		.expect("late review checkpoint should record");
+
+	let error = super::apply_stale_active_release_with_tracker(
+		&tracker,
+		&context.config,
+		&context.workflow,
+		&context.state_store,
+		&diagnostic,
+	)
+	.expect_err("late review checkpoint should block release");
+	let run = context
+		.state_store
+		.run_attempt("run-1626")
+		.expect("run attempt should read")
+		.expect("run should exist");
+
+	assert!(
+		error.to_string().contains("safety inspection changed before apply")
+			|| error.to_string().contains("review authority appeared"),
+		"unexpected release error: {error:?}"
+	);
+	assert_eq!(run.status(), "running");
+	assert!(tracker.label_removals.borrow().is_empty());
+}
