+++ apps/decodex/src/recovery/tests/stale_active/telemetry.rs
@@ -1,660 +1,8 @@
-use tempfile::TempDir;
+mod git_worktree;
+mod goal_recovery;
+mod harness;
+mod runtime_claims;
 
-use crate::{
-	recovery::{
-		RecoveryRuntimeMutationPolicy, STALE_ACTIVE_CLASSIFICATION,
-		tests::{
-			self, GhostLaneTestTracker,
-			stale_active::{self},
-		},
-	},
-	state::StateStore,
-	tracker,
+pub(super) use crate::recovery::{
+	STALE_ACTIVE_BLOCKED_CLASSIFICATION, diagnose_stale_active_issues,
 };
-
-#[test]
-fn stale_active_diagnose_allows_dead_orphan_thread_runtime_telemetry() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let store = StateStore::open_in_memory().expect("state store should open");
-	let workflow = tests::sample_workflow();
-	let active_label = tracker::automation_active_label("pubfi");
-	let queue_label = tracker::automation_queue_label("pubfi");
-	let worktree_path = temp_dir.path().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("In Progress", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let diagnostics = super::diagnose_stale_active_issues(
-		"pubfi",
-		&workflow,
-		temp_dir.path(),
-		&store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.first().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, STALE_ACTIVE_CLASSIFICATION);
-	assert!(diagnostic.recoverable(), "unexpected blockers: {:?}", diagnostic.blockers);
-	assert!(diagnostic.evidence.contains(&String::from("process_not_alive")));
-	assert!(diagnostic.evidence.contains(&String::from("stale_active_control_channel_present")));
-	assert!(
-		diagnostic
-			.evidence
-			.contains(&String::from("only_stale_active_or_failed_control_evidence_present"))
-	);
-}
-
-#[test]
-fn stale_active_diagnose_allows_dead_process_leased_claim_cleanup() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let store = StateStore::open_in_memory().expect("state store should open");
-	let workflow = tests::sample_workflow();
-	let active_label = tracker::automation_active_label("pubfi");
-	let queue_label = tracker::automation_queue_label("pubfi");
-	let worktree_path = temp_dir.path().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("Todo", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
-
-	store
-		.upsert_lease("pubfi", &issue.id, "run-1626", "In Progress")
-		.expect("dead run lease should remain recorded");
-
-	stale_active::append_dead_process_interrupt_control_telemetry(&store, &issue.id);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let diagnostics = super::diagnose_stale_active_issues(
-		"pubfi",
-		&workflow,
-		temp_dir.path(),
-		&store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.first().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, STALE_ACTIVE_CLASSIFICATION);
-	assert!(diagnostic.run_lease);
-	assert!(diagnostic.active_shared_claim);
-	assert!(diagnostic.recoverable(), "unexpected blockers: {:?}", diagnostic.blockers);
-	assert!(diagnostic.evidence.contains(&String::from("stale_run_lease_present")));
-	assert!(diagnostic.evidence.contains(&String::from("stale_active_shared_claim_present")));
-	assert!(
-		diagnostic
-			.evidence
-			.contains(&String::from("only_stale_active_or_failed_control_evidence_present"))
-	);
-}
-
-#[test]
-fn stale_active_diagnose_blocks_dead_marker_for_different_leased_run() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let store = StateStore::open_in_memory().expect("state store should open");
-	let workflow = tests::sample_workflow();
-	let active_label = tracker::automation_active_label("pubfi");
-	let queue_label = tracker::automation_queue_label("pubfi");
-	let worktree_path = temp_dir.path().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("Todo", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
-
-	store
-		.record_run_attempt("run-latest", &issue.id, 2, "running")
-		.expect("latest run should record");
-	store
-		.upsert_lease("pubfi", &issue.id, "run-latest", "In Progress")
-		.expect("dead run lease should remain recorded");
-
-	stale_active::append_dead_process_interrupt_control_telemetry(&store, &issue.id);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let diagnostics = super::diagnose_stale_active_issues(
-		"pubfi",
-		&workflow,
-		temp_dir.path(),
-		&store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.first().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
-	assert!(diagnostic.blockers.contains(&String::from("run_lease_present")));
-	assert!(diagnostic.blockers.contains(&String::from("active_shared_claim_present")));
-	assert!(
-		!diagnostic.evidence.contains(&String::from("stale_run_lease_present")),
-		"mismatched marker must not authorize lease cleanup"
-	);
-	assert!(!diagnostic.recoverable());
-}
-
-#[test]
-fn stale_active_diagnose_blocks_dead_marker_for_same_run_different_attempt() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let store = StateStore::open_in_memory().expect("state store should open");
-	let workflow = tests::sample_workflow();
-	let active_label = tracker::automation_active_label("pubfi");
-	let queue_label = tracker::automation_queue_label("pubfi");
-	let worktree_path = temp_dir.path().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("Todo", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
-
-	store
-		.record_run_attempt("run-1626", &issue.id, 2, "running")
-		.expect("later attempt should record");
-	store
-		.upsert_lease("pubfi", &issue.id, "run-1626", "In Progress")
-		.expect("dead run lease should remain recorded");
-
-	stale_active::append_dead_process_interrupt_control_telemetry(&store, &issue.id);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let diagnostics = super::diagnose_stale_active_issues(
-		"pubfi",
-		&workflow,
-		temp_dir.path(),
-		&store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.first().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
-	assert_eq!(diagnostic.latest_run_id.as_deref(), Some("run-1626"));
-	assert_eq!(diagnostic.latest_attempt_number, Some(2));
-	assert!(diagnostic.blockers.contains(&String::from("run_lease_present")));
-	assert!(diagnostic.blockers.contains(&String::from("active_shared_claim_present")));
-	assert!(
-		!diagnostic.evidence.contains(&String::from("stale_run_lease_present")),
-		"mismatched attempt must not authorize lease cleanup"
-	);
-	assert!(!diagnostic.recoverable());
-}
-
-#[test]
-fn stale_active_diagnose_blocks_external_shared_claim_over_dead_local_lease() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let store = StateStore::open_in_memory().expect("state store should open");
-	let external_store = StateStore::open_in_memory().expect("external store should open");
-	let workflow = tests::sample_workflow();
-	let claim_root = temp_dir.path().join("claims");
-	let active_label = tracker::automation_active_label("pubfi");
-	let queue_label = tracker::automation_queue_label("pubfi");
-	let worktree_path = temp_dir.path().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("Todo", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	store
-		.configure_dispatch_slot_root("pubfi", &claim_root)
-		.expect("store should configure shared claims");
-	external_store
-		.configure_dispatch_slot_root("pubfi", &claim_root)
-		.expect("external store should configure shared claims");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
-
-	store
-		.upsert_lease("pubfi", &issue.id, "run-1626", "In Progress")
-		.expect("dead run lease should remain recorded");
-
-	stale_active::append_dead_process_interrupt_control_telemetry(&store, &issue.id);
-
-	assert!(
-		external_store
-			.try_acquire_lease("pubfi", &issue.id, "run-external", "In Progress")
-			.expect("external claim should attempt"),
-		"external claim should acquire the shared issue lock"
-	);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let diagnostics = super::diagnose_stale_active_issues(
-		"pubfi",
-		&workflow,
-		temp_dir.path(),
-		&store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.first().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
-	assert!(diagnostic.blockers.contains(&String::from("active_shared_claim_present")));
-	assert!(
-		diagnostic.evidence.contains(&String::from("stale_active_claim_identity_mismatch_present"))
-	);
-	assert!(!diagnostic.recoverable());
-}
-
-#[test]
-fn stale_active_diagnose_allows_app_server_no_progress_failure_evidence() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let store = StateStore::open_in_memory().expect("state store should open");
-	let workflow = tests::sample_workflow();
-	let active_label = tracker::automation_active_label("pubfi");
-	let queue_label = tracker::automation_queue_label("pubfi");
-	let worktree_path = temp_dir.path().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("In Progress", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
-	stale_active::append_app_server_no_progress_failure_evidence(&store, &issue.id);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let diagnostics = super::diagnose_stale_active_issues(
-		"pubfi",
-		&workflow,
-		temp_dir.path(),
-		&store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.first().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, STALE_ACTIVE_CLASSIFICATION);
-	assert!(diagnostic.recoverable(), "unexpected blockers: {:?}", diagnostic.blockers);
-	assert!(
-		diagnostic
-			.evidence
-			.contains(&String::from("only_stale_active_or_failed_control_evidence_present"))
-	);
-}
-
-#[test]
-fn stale_active_diagnose_blocks_harness_outcome_with_pr_progress() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let store = StateStore::open_in_memory().expect("state store should open");
-	let workflow = tests::sample_workflow();
-	let active_label = tracker::automation_active_label("pubfi");
-	let queue_label = tracker::automation_queue_label("pubfi");
-	let worktree_path = temp_dir.path().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("In Progress", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
-	stale_active::append_harness_outcome_with_pr_progress(&store, &issue.id);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let diagnostics = super::diagnose_stale_active_issues(
-		"pubfi",
-		&workflow,
-		temp_dir.path(),
-		&store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.first().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
-	assert!(diagnostic.blockers.contains(&String::from("private_progress_evidence_present")));
-	assert!(!diagnostic.recoverable());
-}
-
-#[test]
-fn stale_active_diagnose_blocks_harness_outcome_with_review_progress() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let store = StateStore::open_in_memory().expect("state store should open");
-	let workflow = tests::sample_workflow();
-	let active_label = tracker::automation_active_label("pubfi");
-	let queue_label = tracker::automation_queue_label("pubfi");
-	let worktree_path = temp_dir.path().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("In Progress", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
-	stale_active::append_harness_outcome_with_review_progress(&store, &issue.id);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let diagnostics = super::diagnose_stale_active_issues(
-		"pubfi",
-		&workflow,
-		temp_dir.path(),
-		&store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.first().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
-	assert!(diagnostic.blockers.contains(&String::from("private_progress_evidence_present")));
-	assert!(!diagnostic.recoverable());
-}
-
-#[test]
-fn stale_active_diagnose_blocks_harness_outcome_with_validation_progress() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let store = StateStore::open_in_memory().expect("state store should open");
-	let workflow = tests::sample_workflow();
-	let active_label = tracker::automation_active_label("pubfi");
-	let queue_label = tracker::automation_queue_label("pubfi");
-	let worktree_path = temp_dir.path().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("In Progress", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
-	stale_active::append_harness_outcome_with_validation_progress(&store, &issue.id);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let diagnostics = super::diagnose_stale_active_issues(
-		"pubfi",
-		&workflow,
-		temp_dir.path(),
-		&store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.first().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
-	assert!(diagnostic.blockers.contains(&String::from("private_progress_evidence_present")));
-	assert!(!diagnostic.recoverable());
-}
-
-#[test]
-fn stale_active_diagnose_blocks_no_diff_guardrail_with_delta() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let store = StateStore::open_in_memory().expect("state store should open");
-	let workflow = tests::sample_workflow();
-	let active_label = tracker::automation_active_label("pubfi");
-	let queue_label = tracker::automation_queue_label("pubfi");
-	let worktree_path = temp_dir.path().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("In Progress", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
-	stale_active::append_no_diff_guardrail_event(&store, &issue.id, true, false);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let diagnostics = super::diagnose_stale_active_issues(
-		"pubfi",
-		&workflow,
-		temp_dir.path(),
-		&store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.first().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
-	assert!(diagnostic.blockers.contains(&String::from("private_progress_evidence_present")));
-	assert!(!diagnostic.recoverable());
-}
-
-#[test]
-fn stale_active_diagnose_allows_startup_no_diff_guardrail_without_error_class() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let store = StateStore::open_in_memory().expect("state store should open");
-	let workflow = tests::sample_workflow();
-	let active_label = tracker::automation_active_label("pubfi");
-	let queue_label = tracker::automation_queue_label("pubfi");
-	let worktree_path = temp_dir.path().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("In Progress", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
-	stale_active::append_no_diff_guardrail_event_with_source_error_class(
-		&store, &issue.id, false, false, None,
-	);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let diagnostics = super::diagnose_stale_active_issues(
-		"pubfi",
-		&workflow,
-		temp_dir.path(),
-		&store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.first().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, STALE_ACTIVE_CLASSIFICATION);
-	assert!(diagnostic.recoverable(), "unexpected blockers: {:?}", diagnostic.blockers);
-}
-
-#[test]
-fn stale_active_diagnose_allows_app_server_phase_goal_recovery_telemetry() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let store = StateStore::open_in_memory().expect("state store should open");
-	let workflow = tests::sample_workflow();
-	let active_label = tracker::automation_active_label("pubfi");
-	let queue_label = tracker::automation_queue_label("pubfi");
-	let worktree_path = temp_dir.path().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("In Progress", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
-	stale_active::append_phase_goal_recovery_event(
-		&store,
-		&issue.id,
-		"implement_to_validation_ready",
-		"app_server_dynamic_tool_protocol_failure",
-	);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let diagnostics = super::diagnose_stale_active_issues(
-		"pubfi",
-		&workflow,
-		temp_dir.path(),
-		&store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.first().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, STALE_ACTIVE_CLASSIFICATION);
-	assert!(diagnostic.recoverable(), "unexpected blockers: {:?}", diagnostic.blockers);
-}
-
-#[test]
-fn stale_active_diagnose_blocks_repo_gate_phase_goal_recovery_telemetry() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let store = StateStore::open_in_memory().expect("state store should open");
-	let workflow = tests::sample_workflow();
-	let active_label = tracker::automation_active_label("pubfi");
-	let queue_label = tracker::automation_queue_label("pubfi");
-	let worktree_path = temp_dir.path().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("In Progress", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
-	stale_active::append_phase_goal_recovery_event(
-		&store,
-		&issue.id,
-		"implement_to_validation_ready",
-		"repo_gate_verify_failed",
-	);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let diagnostics = super::diagnose_stale_active_issues(
-		"pubfi",
-		&workflow,
-		temp_dir.path(),
-		&store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.first().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
-	assert!(diagnostic.blockers.contains(&String::from("private_progress_evidence_present")));
-	assert!(!diagnostic.recoverable());
-}
-
-#[test]
-fn stale_active_diagnose_blocks_repair_phase_goal_recovery_telemetry() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let store = StateStore::open_in_memory().expect("state store should open");
-	let workflow = tests::sample_workflow();
-	let active_label = tracker::automation_active_label("pubfi");
-	let queue_label = tracker::automation_queue_label("pubfi");
-	let worktree_path = temp_dir.path().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("In Progress", &[active_label, queue_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	stale_active::seed_dead_orphan_runtime_telemetry(&store, &issue, &worktree_path);
-	stale_active::append_phase_goal_recovery_event(
-		&store,
-		&issue.id,
-		"repair_accepted_review_findings",
-		"app_server_dynamic_tool_protocol_failure",
-	);
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let diagnostics = super::diagnose_stale_active_issues(
-		"pubfi",
-		&workflow,
-		temp_dir.path(),
-		&store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.first().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.classification, super::super::super::STALE_ACTIVE_BLOCKED_CLASSIFICATION);
-	assert!(diagnostic.blockers.contains(&String::from("private_progress_evidence_present")));
-	assert!(!diagnostic.recoverable());
-}
-
-#[test]
-fn stale_active_diagnose_blocks_clean_worktree_with_unmerged_commits() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let store = StateStore::open_in_memory().expect("state store should open");
-	let workflow = tests::sample_workflow();
-	let active_label = tracker::automation_active_label("pubfi");
-	let worktree_path = temp_dir.path().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("Todo", &[active_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	tests::init_git_repo(&worktree_path);
-	tests::run_git(&worktree_path, &["checkout", "-B", "main"]);
-	tests::commit_test_file(&worktree_path, "README.md", "base\n", "base");
-	tests::run_git(&worktree_path, &["checkout", "-b", "x/pubfi-pub-1626"]);
-	tests::commit_test_file(&worktree_path, "source.rs", "fn retained_progress() {}\n", "progress");
-
-	store
-		.record_run_attempt("run-1626", &issue.id, 1, "running")
-		.expect("run attempt should record");
-	store
-		.upsert_worktree(
-			"pubfi",
-			&issue.id,
-			"x/pubfi-pub-1626",
-			&worktree_path.display().to_string(),
-		)
-		.expect("worktree mapping should record");
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let diagnostics = super::diagnose_stale_active_issues(
-		"pubfi",
-		&workflow,
-		temp_dir.path(),
-		&store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.first().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.worktree_state, "unmerged_commits_present");
-	assert!(diagnostic.blockers.contains(&String::from("worktree_unmerged_commits_present")));
-	assert!(
-		diagnostic.next_action.contains("Preserve retained progress")
-			&& diagnostic.next_action.contains("inspect the retained worktree"),
-		"retained worktree blockers should route to retained-progress inspection, got {:?}",
-		diagnostic.next_action
-	);
-	assert!(!diagnostic.recoverable());
-}
-
-#[test]
-fn stale_active_diagnose_blocks_clean_git_worktree_without_default_branch() {
-	let temp_dir = TempDir::new().expect("tempdir should create");
-	let store = StateStore::open_in_memory().expect("state store should open");
-	let workflow = tests::sample_workflow();
-	let active_label = tracker::automation_active_label("pubfi");
-	let worktree_path = temp_dir.path().join("PUB-1626");
-	let mut issue = tests::sample_issue_with_labels("Todo", &[active_label]);
-
-	issue.identifier = String::from("PUB-1626");
-
-	tests::init_git_repo(&worktree_path);
-
-	store
-		.record_run_attempt("run-1626", &issue.id, 1, "running")
-		.expect("run attempt should record");
-	store
-		.upsert_worktree(
-			"pubfi",
-			&issue.id,
-			"x/pubfi-pub-1626",
-			&worktree_path.display().to_string(),
-		)
-		.expect("worktree mapping should record");
-
-	let tracker = GhostLaneTestTracker::with_issues(vec![issue]);
-	let diagnostics = super::diagnose_stale_active_issues(
-		"pubfi",
-		&workflow,
-		temp_dir.path(),
-		&store,
-		&tracker,
-		Some("PUB-1626"),
-		RecoveryRuntimeMutationPolicy::ReadOnly,
-	)
-	.expect("stale active diagnosis should run");
-	let diagnostic = diagnostics.first().expect("diagnostic should exist");
-
-	assert_eq!(diagnostic.worktree_state, "default_branch_unavailable");
-	assert!(diagnostic.blockers.contains(&String::from("worktree_default_branch_unavailable")));
-	assert!(!diagnostic.recoverable());
-}
